requirements.txt
Pillow==11.1.0
python-bidi==0.6.3
matplotlib
pandas
pytesseract
werkzeug
flask
